New Concept: AppleNotesScriptLayer
