--- Apple Notes Exporter/Apple Notes Exporter/AppleScript.swift
@@ -1,40 +1,0 @@
-//
-//  AppleScript.swift
-//  Apple Notes Exporter
-//
-//  Created by Konstantin Zaremski on 7/11/23.
-//
-
-import Foundation
-
-extension NSAppleEventDescriptor {
-    func toStringArray() -> [String] {
-        guard let listDescriptor = self.coerce(toDescriptorType: typeAEList) else {
-            return []
-        }
-        
-        return (0..<listDescriptor.numberOfItems)
-            .compactMap { listDescriptor.atIndex($0 + 1)?.stringValue }
-    }
-}
-
-class AppleScript {
-    /**
-     Runs an AppleScript script string with an expected string array result.
-     */
-    static func stringArray(script: String) -> [String] {
-        // Create the new NSAppleScript instance
-        if let scriptObject = NSAppleScript(source: script) {
-            // Error dictionary
-            var errorDict: NSDictionary? = nil
-            // Execute the script, adding to the errorDict if there are errors
-            let resultDescriptor = scriptObject.executeAndReturnError(&errorDict)
-            // If there are no errors, return the resultDescriptor after converting it to a string array
-            if errorDict == nil {
-                return resultDescriptor.toStringArray()
-            }
-        }
-        // Return an empty string if no result
-        return []
-    }
-}
