+++ Apple Notes Exporter/Apple Notes Exporter/ExportView.swift
@@ -1,0 +1,18 @@
+//
+//  ExportView.swift
+//  Apple Notes Exporter
+//
+//  Created by Konstantin Zaremski on 5/13/24.
+//
+
+import SwiftUI
+
+struct ExportView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ExportView()
+}
